refactor: move table creation and log summary endpoint
