+++ backend/main.py
@@ -1,17 +1,10 @@
-import os
-
-import openai
 from dotenv import load_dotenv
-from fastapi import FastAPI, File, HTTPException, UploadFile
+from fastapi import FastAPI
 
 from app.api.v1.endpoints.api import api_router
-from app.database import Base, engine
+from app.api.v1.endpoints import log_summary
 
 load_dotenv()
-
-# Создаем таблицы в БД (для первого запуска)
-# В продакшене лучше использовать Alembic
-Base.metadata.create_all(bind=engine)
 
 app = FastAPI(
     title="LeadConverter Pro API",
@@ -21,42 +14,8 @@ app = FastAPI(
 
 # Подключаем роутеры
 app.include_router(api_router, prefix="/api/v1")
+app.include_router(log_summary.router)
 
 @app.get("/")
 def read_root():
     return {"message": "Welcome to LeadConverter Pro API"}
-
-
-@app.post("/summarize-log")
-async def summarize_log(file: UploadFile = File(...)):
-    try:
-        content_bytes = await file.read()
-        log_text = content_bytes.decode("utf-8")
-    except Exception as exc:  # pragma: no cover - simple validation
-        raise HTTPException(status_code=400, detail="Unable to read uploaded file") from exc
-
-    api_key = os.getenv("OPENAI_API_KEY")
-    if not api_key:
-        raise HTTPException(status_code=500, detail="OpenAI API key not configured")
-
-    openai.api_key = api_key
-    prompt = (
-        "Summarize the following application log entries in a concise manner:\n" + log_text
-    )
-
-    try:
-        response = openai.ChatCompletion.create(
-            model="gpt-3.5-turbo",
-            messages=[
-                {
-                    "role": "system",
-                    "content": "You are a helpful assistant that summarizes logs.",
-                },
-                {"role": "user", "content": prompt},
-            ],
-        )
-    except Exception as exc:  # pragma: no cover - API errors
-        raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    summary = response.choices[0].message["content"].strip()
-    return {"summary": summary}
